Worked on the roles page.

+++ views/admin/roles/overview.php
@@ -13,6 +13,13 @@
         <div class="p-6 text-medium text-stone-50 rounded w-full min-h-screen">
             <h3 class="text-lg font-bold text-stone-100 text-white mb-2">Rollen beheren</h3>
             <p class="mb-2"></p>
+            <form method="POST" action="<?php echo $newUrl ?>">
+                <button
+                    class=" mt-6 shadow bg-stone-700 hover:bg-stone-300 focus:shadow-outline focus:outline-none text-white font-bold py-2 px-4 rounded"
+                    type="submit">
+                    Rol toevoegen
+                </button>
+            </form>
             <br>
             <table class="table-auto w-full bg-gray-800 text-white">
                 <thead>
@@ -23,32 +30,25 @@
                     </tr>
                 </thead>
                 <tbody>
-                    <tr class="even:bg-stone-900 odd:bg-stone-950">
-                        <td class="px-4 py-2">
-                            <a href="/admin/roles/delete">
-                                <img src="/images/trash.svg" alt="Trash" />
-                            </a>
-                        </td>
-                        <td class="px-4 py-2">
-                            <a class="underline" href="/admin/users/edit">
-                                Student
-                            </a>
-                        </td>
-                        <td class="px-4 py-2">10</td>
-                    </tr>
-                    <tr class="even:bg-stone-900 odd:bg-stone-950">
-                        <td class="px-4 py-2">
-                            <a href="/admin/roles/delete">
-                                <img src="/images/trash.svg" alt="Trash" />
-                            </a>
-                        </td>
-                        <td class="px-4 py-2">
-                            <a class="underline" href="/admin/users/edit">
-                                Docent
-                            </a>
-                        </td>
-                        <td class="px-4 py-2">20</td>
-                    </tr>
+                    <?php
+                    foreach ($roles as $role) {
+                        ?>
+                        <tr class="even:bg-stone-900 odd:bg-stone-950">
+                            <td class="px-4 py-2">
+                                <a href="/admin/roles/delete?id=<?php echo $role["id"]; ?>" onclick="return confirm('Weet je zeker dat je deze roll wil verwijderen?');">
+                                    <img src=" /images/trash.svg" alt="Trash" />
+                                </a>
+                            </td>
+                            <td class="px-4 py-2">
+                                <a class="underline" href="/admin/roles/edit?id=<?php echo $role["id"]; ?>">
+                                    <?php echo $role["name"]; ?>
+                                </a>
+                            </td>
+                            <td class="px-4 py-2"> <?php echo $role["level"]; ?></td>
+                        </tr>
+                        <?php
+                    }
+                    ?>
                 </tbody>
             </table>
         </div>
